+++ package.json
@@ -38,19 +38,14 @@
     "watch": "vitest"
   },
   "devDependencies": {
-    "@bufbuild/buf": "^1.28.1",
     "@bufbuild/protobuf": "^1.6.0",
-    "@bufbuild/protoc-gen-es": "^1.6.0",
     "@codemirror/commands": "^6.3.0",
     "@codemirror/language": "^6.9.2",
     "@codemirror/lint": "^6.4.2",
     "@codemirror/state": "^6.3.1",
     "@codemirror/view": "^6.21.4",
     "@lezer/highlight": "^1.1.6",
-    "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-icons": "^1.3.0",
-    "@radix-ui/react-toast": "^1.1.5",
-    "@radix-ui/react-tooltip": "^1.0.7",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
     "@typescript-eslint/eslint-plugin": "^7.9.0",
