+++ src/web/dusa.css
@@ -148,7 +148,6 @@ button {
   display: none;
 }
 
-
 button:disabled {
   opacity: 0.5;
 }
